feat: add PageSkeleton component, cron service, and UI improvements
- Add PageSkeleton loading component with CSS animations
- Add cron_service for background scheduling
- Refactor TaxTypeDropdown accessibility (aria-expanded, aria-selected)
- Move inline styles to ap-invoice.css and page-skeleton.css
- Improve usage_service with expanded functionality
- Update index.html with enhanced metadata
- Fix auth session handling and router updates

+++ backend/app/auth/session.py
@@ -42,7 +42,7 @@ def create_session_jwt(
     carmen_user_id: str,
     username: str,
     secret: str,
-    ttl_hours: int = 8,
+    ttl_hours: float = 8,
     carmen_uri: str = "",
     bu: str = "",
 ) -> str:
